無聊/requirements.txt
# 台指期 Discord 機器人即時預警系統 - 依賴套件

# 核心資料處理
pandas>=1.5.0
numpy>=1.21.0
TA-Lib>=0.4.24

# 非同步 HTTP 請求
aiohttp>=3.8.0

# 設定檔處理
PyYAML>=6.0

# 日誌處理
colorlog>=6.7.0

# 可選：IB API 連接（如果需要使用 IB 資料源）
# ib_insync>=0.9.70

# 可選：Discord.py（如果需要更複雜的 Bot 功能）
# discord.py>=2.0.0 